requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
pypdf
pymupdf
sentence-transformers
langchain-huggingface
faiss-cpu
chromadb
langchain-groq
langchain-openai
python-dotenv
openai
gradio
beautifulsoup4
requests
chardet
